Renamne til RisikoscenarioSannsynlighetField
