+++ apps/frontend/src/components/risikoscenario/edit/RisikoscenarioModalForm.tsx
@@ -5,7 +5,7 @@ import { mapRisikoscenarioToFormValue } from '../../../api/RisikoscenarioApi'
 import { IRisikoscenario } from '../../../constants'
 import { RisikoscenarioKonsekvensnivaaField } from './RisikoscenarioKonsekvensnivaaField/RisikoscenarioKonsekvensnivaaField'
 import { RisikoscenarioNavnBeskrivelseField } from './RisikoscenarioNavnBeskrivelseField/RisikoscenarioNavnBeskrivelseField'
-import { RisikoscenarioSannsynlighet } from './RisikoscenarioSannsynlighet/RisikoscenarioSannsynlighet'
+import { RisikoscenarioSannsynlighetField } from './RisikoscenarioSannsynlighetField/RisikoscenarioSannsynlighetField'
 import { risikoscenarioCreateValidation } from './RisikoscenarioSchemaValidation'
 
 type TProps = {
@@ -45,7 +45,7 @@ export const RisikoscenarioModalForm: FunctionComponent<TProps> = ({
             <Modal.Body>
               <RisikoscenarioNavnBeskrivelseField />
 
-              <RisikoscenarioSannsynlighet />
+              <RisikoscenarioSannsynlighetField />
 
               <RisikoscenarioKonsekvensnivaaField />
             </Modal.Body>
